requirements.txt
flask
twilio
langchain
openai
pandas
gspread
oauth2client
python-dotenv
langchain_openai